@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class Avatar extends StatelessWidget {
-  const Avatar({Key key, this.photoUrl, this.radius = 50}) : super(key: key);
+  const Avatar({Key? key, required this.photoUrl, this.radius = 50}) : super(key: key);
   final String photoUrl;
   final double radius;
 
