@@ -1,5 +1,5 @@
-class Split {
-  const Split({this.km, this.pace, this.elev});
+class Splits {
+  const Splits({required this.km, required this.pace, required this.elev});
   final int km;
   final Duration pace;
   final int elev;
