@@ -1,5 +1,5 @@
 class User {
-  const User({this.displayName, this.username, this.photoUrl});
+  const User({required this.displayName, required this.username, required this.photoUrl});
   final String displayName;
   final String username;
   final String photoUrl;
