@@ -20,7 +20,7 @@ class NestedScrollViewPage extends StatelessWidget {
               // This is not necessary if the "headerSliverBuilder" only builds
               // widgets that do not overlap the next sliver.
               handle: NestedScrollView.sliverOverlapAbsorberHandleFor(context),
-              child: SliverAppBar(
+              sliver: SliverAppBar(
                 title: const Text('Books'), // This is the title in the app bar.
                 pinned: true,
                 expandedHeight: 150.0,
